make vertex/pdqvertexfactory an inner class
